Wrote simple test for GitHubAPI.
