+++ src/main/resources/reference.conf
@@ -22,9 +22,7 @@ codefeedr {
   input {
     github {
       # Api keys for querying GitHub events
-      apikey {
-        0 = ""
-      }
+      apikeys = [""]
 
       # Name of the collection to use for events
       events_collection = "github_events"
