+++ src/test/resources/reference.conf
@@ -37,9 +37,7 @@ codefeedr {
   input {
     github {
       # Api key for querying GitHub events
-      apikey {
-        0 = ""
-      }
+      apikeys = ["0", "1", "2"]
 
       # Name of the collection to use for events
       events_collection = "github_events"
